+++ lib/easy_to_dictate_numbers.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-# lib/easy_to_dictate_numbers.rb
-require_relative 'easy_to_dictate_numbers/version'
-require_relative 'easy_to_dictate_numbers/extend_primitives'
-require_relative 'easy_to_dictate_numbers/exceptions'
-require_relative 'easy_to_dictate_numbers/dictable_number'